DynamoDB - Support dots in ConditionExpressions
